add lychee link-checking workflows, config, script
Lychee[0] is a fast link checker that supports .md. This commit adds the
tool to the flake, provides a helper script & pre-commit check, for
local dev, and adds two CI jobs.

The first CI job runs for all pushes/PRs and checks links with a failing
status if there are any bad ones.

The second CI job runs as a cron job and doesn't fail for bad links, but
does create repo issues to flag them as needing to be fixed.

[0]: https://github.com/lycheeverse/lychee