initial pdoc API docs & type checking stubs
After trying a few different "automated" approaches to generating API
documentation for the Python APIs I've decided the best results are
obtained with manual curation of `.pyi` stub files. Le sigh.

Keeping these in-sync and error free will be a challenge, but it's the
best way to carefully craft documentation for the most critical user
facing surface of the client.

The general idea is to:

* Write `.pyi` stub files for each of the important user-facing Python
  APIs/types/etc.
* Configure type checking tools to use these stubs (TODO: pyright)
* When we want to render API docs, rename each `.pyi` to `py` and run it
  through the pdoc tool, then rename each back to `.pyi`.
* Publish the resulting generated docs as part of the site.
* ???
* Profit

There's a ton more docs to write, but let's get the mechanics worked out
before going for exhaustive coverage.