remove python-typings and .pyi stubs
It isn't possible to meaningfully generate these automatically from the
PyO3 Rust definitions. Keeping them up-to-date manually will suck. I was
planning to swallow that bitter pill with the hope it would at least
make generating Python API docs easier. Turns out it doesn't: we can do
that at runtime without needing stubs.

So: let's remove the stubs for now. We can revisit this in the future
if/when it becomes possible to generate stubs from PyO3.