deploy both api docs and user-guide
This removes the checked-in generated api-docs in favour of generating
them on the fly when needed, and deploying a copy from `main` to the
GitHub pages site. This avoids having to verify the checked in copy are
"up to date" with new changes, or tooling version mismatches.

The existing user-guide content that was at the root of the site is now
moved to a subdir. README and other links should be updated to point
directly at the guide. The api-docs are in their own subdir alongside.
A crummy hardcoded index.html is added in case anyone lands at the root
of the site.