python: catch warnings to avoid stdout corruption
The `mudpuppy` Python module is a good place to set up a custom Python
`warnings.showwarning` handler to redirect warnings to `logging` and the
active session's output buffer.

The default behaviour is to print the warning to stdout, but since we'r
drawing a TUI there it will corrupt the interface.

This is particularly important because common mistakes like forgetting
to await a coroutine will provoke this warning path. We both need to
avoid corrupting the UI in that case, but also surface the error to the
user.