user-guide: cross-link to API docs, refine content
Throughout, try to link to relevant parts of the API documentation now
that it exists. No doubt there are more places to do this, but here's
a first pass.

Along the way, a few fixes, content tweaks, and more coverage. Also
sketch out a TBD section for a more involved tutorial.