python: skip first tick of timers
The Tokio timer implementation's `tick()` function always yields
immediately on the first poll. This is counter to user expectation
because it means when you add a timer with a duration, it will run
immediately, not after the duration expires.

Work-around this with a quick'n'cheap bool check to skip the first
iteration without affecting the tick count.