python-examples: persist gmcp status data pre-reload
When a `/reload` occurs we need to persist the data we've collected from
GMCP, and then read it back again when the module is initialized anew.
The MUD won't resend earlier GMCP data unprompted so otherwise it
appears as though the status area is broken and missing data.

There's probably room for a Mudpuppy primitive for managing this kind of
state but in the meantime do the simple thing: write data to a JSON file
pre-reload, read it & remove it on init.