expose Input directly to Python scripts
This means moving from `Input` to `Py<Input>` in the client struct, and
acquiring the GIL for use from Rust. This was my original intent
(`Input` is already a pyclass and has stubs!) and simplifies the
`mudpuppy_core` API from the Python side. The `history.py` built-in
module shows the improvement nicely.