input area markup, spellcheck example
Adds new `Input` APIs for decorating the to-be-sent queued input text
with ANSI colour. The markup text is interleaved when displaying the
text in the TUI, but is not transmitted to the MUD.

An example is provided that uses a key event handler to provide
highlighting of known slash commands, and if the `spylls` Python package
is installed, spellchecking of input text (en_US only for the example).