config: add command_prefix setting
This allows replacing `/` for slash commands like `/connect` with
a custom prefix like `!!` for `!!conect`.

Along the way, rejig config settings slightly to:

A) allow giving `None` to resolve a global setting. This is mostly
convenience for the headless UI that wants to know the command prefix in
the global context where no sessions are created yet.

B) rename `resolve_setting` to `resolve_extra_setting` since it
specifically looks in the extra config element.