await new sesh handlers, drain autoconnect command
Doing this in parallel meant that it was unpredictable if a new session
handler's subsequently registered event handlers would be registered in
time to actually catch the events produced. Impl is a bit clunky, but
probably good enough (TM) for now.

Similarly, the auto-connect logic might have put commands into the queue
and we need to drain+execute those before the event loop.